+++ playground/setupTests.js
@@ -1,0 +1,5 @@
+import { setDomPreviewContext } from "dom-preview";
+
+beforeEach(() => {
+  setDomPreviewContext(expect.getState().currentTestName);
+});